refactor(product): split model selector menu
